requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
psutil
sympy
wikipedia
pint
newsapi-python
google-api-python-client
pytz
google-auth-oauthlib
google-auth-httplib2